requirements.txt
networkx
numpy
fastapi
uvicorn
playwright
requests
scipy
websocket-client
web3
python-dotenv
